Fix ptp and move the prometheus test
